Create Bank Queue_Two Queues
There are two queues.  One for simple transactions and one for complex transactions.  When a customer arrives, they join the appropriate queue. One teller (a relatively new hire) can handle only simple transactions, so is only able to accept customers from the simple transaction queue.  The other two tellers are more experienced and can handle both simple and complex transactions.  When an experienced teller becomes free and both queues are non-empty the teller randomly chooses to take the next customer from either queue.  If only one of the queues has customers in it, then the teller accepts the next customer from that queue. If there are no customers in either queue, the teller becomes idle. When a customer arrives to find the queues empty and one 
or more tellers available, they randomly select a teller (subject to constraints on the type of service required, i.e. a customer with a complex transaction cannot be served by the new hire.)